[CfgEditor] Remove HtmlBuilder
`CfgHtmlBuilder` is class but used just as function.
Furthermore, since original reason to separate `CfgHtmlBuilder` was because of unit test,
it's time to rollback.

ONE-vscode-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>
